@@ -49,7 +49,7 @@
     "@koa/cors": "3.3.0",
     "@koa/multer": "3.0.0",
     "@koa/router": "10.1.1",
-    "@monaco-editor/react": "4.4.2",
+    "@monaco-editor/react": "4.4.4",
     "ads-client": "1.13.1",
     "axios": "0.26.1",
     "basic-auth": "2.0.1",
@@ -86,14 +86,14 @@
     "pino-pretty": "7.6.1",
     "prop-types": "15.8.1",
     "proxy-agent": "5.0.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-flow-renderer": "10.0.8",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-flow-renderer": "10.2.2",
     "react-icons": "4.3.1",
     "react-json-view": "1.21.3",
     "react-router": "6.3.0",
     "react-router-dom": "6.3.0",
-    "reactstrap": "9.0.1",
+    "reactstrap": "9.0.2",
     "real-require": "git://github.com/burgerni10/real-require.git",
     "sprintf-js": "1.1.2",
     "sqlite": "4.0.25",
