@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import testConfig from '../../../tests/testConfig'
 import Caching from './Caching.jsx'
 
@@ -15,32 +15,37 @@ mockMath.random = () => 1
 global.Math = mockMath
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('Caching', () => {
   test('check Caching', () => {
     act(() => {
-      ReactDOM.render(<Caching
+      root.render(<Caching
         caching={testConfig.engine.caching}
         onChange={() => (1)}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check change cacheFolder', () => {
     act(() => {
-      ReactDOM.render(<Caching
+      root.render(<Caching
         caching={testConfig.engine.caching}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.caching.cacheFolder'), { target: { value: './newFolder' } })
     expect(onChange).toBeCalledWith('engine.caching.cacheFolder', './newFolder', null)
@@ -48,10 +53,10 @@ describe('Caching', () => {
   })
   test('check change archiveFolder', () => {
     act(() => {
-      ReactDOM.render(<Caching
+      root.render(<Caching
         caching={testConfig.engine.caching}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.caching.archive.archiveFolder'), { target: { value: './newArchiveFolder' } })
     expect(onChange).toBeCalledWith('engine.caching.archive.archiveFolder', './newArchiveFolder', null)
@@ -59,10 +64,10 @@ describe('Caching', () => {
   })
   test('check change archiveMode to "archive"', () => {
     act(() => {
-      ReactDOM.render(<Caching
+      root.render(<Caching
         caching={testConfig.engine.caching}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.caching.archive.enabled'), { target: { checked: true } })
     expect(onChange).toBeCalledWith('engine.caching.archive.enabled', true, null)
