@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import testConfig from '../../../tests/testConfig'
 import ExternalSources from './ExternalSources.jsx'
 
@@ -17,30 +17,35 @@ const dispatchNewConfig = jest.fn()
 React.useContext = jest.fn().mockReturnValue({ dispatchNewConfig })
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('ExternalSources', () => {
   test('check ExternalSources', () => {
     act(() => {
-      ReactDOM.render(<ExternalSources
+      root.render(<ExternalSources
         externalSources={testConfig.engine.externalSources}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check change line', () => {
     act(() => {
-      ReactDOM.render(<ExternalSources
+      root.render(<ExternalSources
         externalSources={testConfig.engine.externalSources}
-      />, container)
+      />)
     })
     Simulate.change(document.querySelector('td input'), { target: { value: 'newSource' } })
     expect(dispatchNewConfig).toBeCalledWith({ type: 'update', name: 'engine.externalSources.0', value: 'newSource', validity: null })
@@ -48,20 +53,26 @@ describe('ExternalSources', () => {
   })
   test('check delete ExternalSources', () => {
     act(() => {
-      ReactDOM.render(<ExternalSources
+      root.render(<ExternalSources
         externalSources={testConfig.engine.externalSources}
-      />, container)
+      />)
     })
-    Simulate.click(document.querySelector('td path'))
-    Simulate.click(document.getElementsByClassName('btn btn-primary')[0])
+    act(() => {
+      Simulate.click(document.querySelector('td path'))
+    })
+
+    act(() => {
+      Simulate.click(document.getElementsByClassName('btn btn-primary')[0])
+    })
+
     expect(dispatchNewConfig).toBeCalledWith({ type: 'deleteRow', name: 'engine.externalSources.0' })
     expect(container).toMatchSnapshot()
   })
   test('check add line', () => {
     act(() => {
-      ReactDOM.render(<ExternalSources
+      root.render(<ExternalSources
         externalSources={testConfig.engine.externalSources}
-      />, container)
+      />)
     })
     Simulate.click(document.querySelector('th path'))
     expect(dispatchNewConfig).toBeCalledWith({ type: 'addRow', name: 'engine.externalSources', value: '' })
