@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import testConfig from '../../../tests/testConfig'
 import HttpRequest from './HttpRequest.jsx'
 
@@ -15,32 +15,37 @@ mockMath.random = () => 1
 global.Math = mockMath
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('HttpRequest', () => {
   test('check HttpRequest', () => {
     act(() => {
-      ReactDOM.render(<HttpRequest
+      root.render(<HttpRequest
         httpRequest={testConfig.engine.httpRequest}
         onChange={() => (1)}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check change stack to "request"', () => {
     act(() => {
-      ReactDOM.render(<HttpRequest
+      root.render(<HttpRequest
         httpRequest={testConfig.engine.httpRequest}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.httpRequest.stack'), { target: { value: 'axios', selectedIndex: 0 } })
     expect(onChange).toBeCalledWith('engine.httpRequest.stack', 'axios', null, null)
@@ -48,10 +53,10 @@ describe('HttpRequest', () => {
   })
   test('check change timeout', () => {
     act(() => {
-      ReactDOM.render(<HttpRequest
+      root.render(<HttpRequest
         httpRequest={testConfig.engine.httpRequest}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.httpRequest.timeout'), { target: { value: 60 } })
     expect(onChange).toBeCalledWith('engine.httpRequest.timeout', 60, null)
