@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import testConfig from '../../../../tests/testConfig'
 import ConsoleLogging from './ConsoleLogging.jsx'
 
@@ -15,32 +15,37 @@ mockMath.random = () => 1
 global.Math = mockMath
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('ConsoleLogging', () => {
   test('check ConsoleLogging', () => {
     act(() => {
-      ReactDOM.render(<ConsoleLogging
+      root.render(<ConsoleLogging
         logParameters={testConfig.engine.logParameters.consoleLog}
         onChange={() => (1)}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check change console level to "warning"', () => {
     act(() => {
-      ReactDOM.render(<ConsoleLogging
+      root.render(<ConsoleLogging
         logParameters={testConfig.engine.logParameters.consoleLog}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.logParameters.consoleLog.level'), { target: { value: 'warning', selectedIndex: 3 } })
     expect(onChange).toBeCalledWith('engine.logParameters.consoleLog.level', 'warning', null, null)
