@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import testConfig from '../../../../tests/testConfig'
 import LokiLogging from './LokiLogging.jsx'
 
@@ -15,32 +15,37 @@ mockMath.random = () => 1
 global.Math = mockMath
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('LokiLogging', () => {
   test('check LokiLogging', () => {
     act(() => {
-      ReactDOM.render(<LokiLogging
+      root.render(<LokiLogging
         logParameters={testConfig.engine.logParameters.lokiLog}
         onChange={() => (1)}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check change lokiLog level to "warning"', () => {
     act(() => {
-      ReactDOM.render(<LokiLogging
+      root.render(<LokiLogging
         logParameters={testConfig.engine.logParameters.lokiLog}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.logParameters.lokiLog.level'), {
       target: {
@@ -53,10 +58,10 @@ describe('LokiLogging', () => {
   })
   test('check change lokiLog address name to "http://test.loki.fr"', () => {
     act(() => {
-      ReactDOM.render(<LokiLogging
+      root.render(<LokiLogging
         logParameters={testConfig.engine.logParameters.lokiLog}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.logParameters.lokiLog.lokiAddress'), { target: { value: 'http://test.loki.fr' } })
     expect(onChange).toBeCalledWith('engine.logParameters.lokiLog.lokiAddress', 'http://test.loki.fr', null)
@@ -64,10 +69,10 @@ describe('LokiLogging', () => {
   })
   test('check change lokiLog interval to 10', () => {
     act(() => {
-      ReactDOM.render(<LokiLogging
+      root.render(<LokiLogging
         logParameters={testConfig.engine.logParameters.lokiLog}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.logParameters.lokiLog.interval'), { target: { value: 10 } })
     expect(onChange).toBeCalledWith('engine.logParameters.lokiLog.interval', 10, null)
