@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import testConfig from '../../../../tests/testConfig'
 import SqliteLogging from './SqliteLogging.jsx'
 
@@ -15,32 +15,37 @@ mockMath.random = () => 1
 global.Math = mockMath
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('SqliteLogging', () => {
   test('check SqliteLogging', () => {
     act(() => {
-      ReactDOM.render(<SqliteLogging
+      root.render(<SqliteLogging
         logParameters={testConfig.engine.logParameters.sqliteLog}
         onChange={() => (1)}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check change sqliteLevel to "warning"', () => {
     act(() => {
-      ReactDOM.render(<SqliteLogging
+      root.render(<SqliteLogging
         logParameters={testConfig.engine.logParameters.sqliteLog}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.logParameters.sqliteLog.level'), { target: { value: 'warning', selectedIndex: 4 } })
     expect(onChange).toBeCalledWith('engine.logParameters.sqliteLog.level', 'error', null, null)
@@ -48,10 +53,10 @@ describe('SqliteLogging', () => {
   })
   test('check change sqliteLog fileName to "new_sqliteFilename"', () => {
     act(() => {
-      ReactDOM.render(<SqliteLogging
+      root.render(<SqliteLogging
         logParameters={testConfig.engine.logParameters.sqliteLog}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.logParameters.sqliteLog.fileName'), { target: { value: 'new_sqliteFilename' } })
     expect(onChange).toBeCalledWith('engine.logParameters.sqliteLog.fileName', 'new_sqliteFilename', null)
@@ -59,10 +64,10 @@ describe('SqliteLogging', () => {
   })
   test('check change sqliteLog maxNumberOfLogs to 10000000', () => {
     act(() => {
-      ReactDOM.render(<SqliteLogging
+      root.render(<SqliteLogging
         logParameters={testConfig.engine.logParameters.sqliteLog}
         onChange={onChange}
-      />, container)
+      />)
     })
     Simulate.change(document.getElementById('engine.logParameters.sqliteLog.maxNumberOfLogs'), { target: { value: 10000000 } })
     expect(onChange).toBeCalledWith('engine.logParameters.sqliteLog.maxNumberOfLogs', 10000000, null)
