@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import newConfig from '../../../../tests/testHistoryConfig'
 import HistoryQueryForm from './HistoryQueryForm.jsx'
 
@@ -22,25 +22,29 @@ mockMath.random = () => 1
 global.Math = mockMath
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('HistoryQueryForm', () => {
   newConfig.forEach((historyQuery) => {
     test(`check HistoryQueryForm with data: ${historyQuery.name}`, () => {
       act(() => {
-        ReactDOM.render(
+        root.render(
           <HistoryQueryForm
             query={historyQuery}
           />,
-          container,
         )
       })
       expect(container).toMatchSnapshot()
@@ -49,7 +53,7 @@ describe('HistoryQueryForm', () => {
 
   test('check HistoryQueryForm with empty application', () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <HistoryQueryForm
           query={{
             id: 'unique-id',
@@ -63,7 +67,6 @@ describe('HistoryQueryForm', () => {
             settings: {},
           }}
         />,
-        container,
       )
     })
     expect(container).toMatchSnapshot()
