@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import EngineMenu from './EngineMenu.jsx'
 
 import activeConfig from '../../../tests/testConfig'
@@ -19,15 +19,19 @@ const restartFunction = jest.fn()
 const shutdownFunction = jest.fn()
 
 let container
-
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
   jest.clearAllMocks()
 })
 
@@ -35,9 +39,8 @@ React.useContext = jest.fn().mockReturnValue({ activeConfig })
 describe('EngineMenu', () => {
   test('display EngineMenu page based on config', async () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />,
-        container,
       )
     })
     expect(container).toMatchSnapshot()
@@ -45,69 +48,97 @@ describe('EngineMenu', () => {
 
   test('check restart engine', () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />,
-        container,
       )
     })
-    Simulate.click(document.getElementById('dropdown-toggle'))
-    Simulate.click(document.getElementById('restart'))
-    Simulate.click(document.getElementById('confirm'))
+    act(() => (
+      Simulate.click(document.getElementById('dropdown-toggle'))
+    ))
+    act(() => (
+      Simulate.click(document.getElementById('restart'))
+    ))
+    act(() => (
+      Simulate.click(document.getElementById('confirm'))
+    ))
     expect(restartFunction).toBeCalled()
     expect(container).toMatchSnapshot()
   })
 
   test('check cancel restart', () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />,
-        container,
       )
     })
-    Simulate.click(document.getElementById('dropdown-toggle'))
-    Simulate.click(document.getElementById('restart'))
-    Simulate.click(document.getElementById('cancel'))
+    act(() => (
+      Simulate.click(document.getElementById('dropdown-toggle'))
+    ))
+    act(() => (
+      Simulate.click(document.getElementById('restart'))
+    ))
+    act(() => (
+      Simulate.click(document.getElementById('cancel'))
+    ))
+
     expect(restartFunction).not.toBeCalled()
     expect(container).toMatchSnapshot()
   })
 
   test('check shutdown engine', () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />,
-        container,
       )
     })
-    Simulate.click(document.getElementById('dropdown-toggle'))
-    Simulate.click(document.getElementById('shutdown'))
-    Simulate.click(document.getElementById('confirm'))
-    // expect(shutdownFunction).toBeCalled()
+    act(() => (
+      Simulate.click(document.getElementById('dropdown-toggle'))
+    ))
+    act(() => (
+      Simulate.click(document.getElementById('shutdown'))
+    ))
+    act(() => (
+      Simulate.click(document.getElementById('confirm'))
+    ))
+
     expect(container).toMatchSnapshot()
   })
 
   test('check cancel shutdown', () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />,
-        container,
       )
     })
-    Simulate.click(document.getElementById('dropdown-toggle'))
-    Simulate.click(document.getElementById('shutdown'))
-    Simulate.click(document.getElementById('cancel'))
+    act(() => (
+      Simulate.click(document.getElementById('dropdown-toggle'))
+    ))
+    act(() => (
+      Simulate.click(document.getElementById('shutdown'))
+    ))
+    act(() => (
+      Simulate.click(document.getElementById('cancel'))
+    ))
+
     expect(shutdownFunction).not.toBeCalled()
     expect(container).toMatchSnapshot()
   })
 
   test('go to engine settings', () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />,
-        container,
       )
     })
-    Simulate.click(document.getElementById('dropdown-toggle'))
-    Simulate.click(document.getElementById('oi-settings'))
+
+    act(() => (
+      Simulate.click(document.getElementById('dropdown-toggle'))
+    ))
+
+    act(() => (
+      Simulate.click(document.getElementById('oi-settings'))
+    ))
+
     expect(mockNavigate).toBeCalledWith('/engine')
     expect(container).toMatchSnapshot()
   })
