@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import NewNorth from './NewNorth.jsx'
 import EngineMenu from './EngineMenu.jsx'
 
@@ -17,7 +17,9 @@ const mockNavigate = jest.fn()
 jest.mock('react-router-dom', () => ({ useNavigate: () => mockNavigate }))
 
 let container
-
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   React.useContext = jest.fn().mockReturnValue({
     newConfig,
@@ -45,22 +47,22 @@ beforeEach(() => {
       },
     ],
   })
-
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('NewNorth', () => {
   test('display NewNorth page based on config', async () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <NewNorth modal={false} toggle={() => false} />,
-        container,
       )
     })
     expect(container).toMatchSnapshot()
@@ -93,9 +95,8 @@ describe('NewNorth', () => {
     })
 
     act(() => {
-      ReactDOM.render(
+      root.render(
         <NewNorth modal={false} toggle={() => false} />,
-        container,
       )
     })
     expect(container).toMatchSnapshot()
@@ -128,9 +129,8 @@ describe('NewNorth', () => {
     })
 
     act(() => {
-      ReactDOM.render(
+      root.render(
         <NewNorth modal={false} toggle={() => false} />,
-        container,
       )
     })
     expect(container).toMatchSnapshot()
@@ -163,10 +163,7 @@ describe('NewNorth', () => {
     })
 
     act(() => {
-      ReactDOM.render(
-        <NewNorth modal={false} toggle={() => false} />,
-        container,
-      )
+      root.render(<NewNorth modal={false} toggle={() => false} />)
     })
     expect(container).toMatchSnapshot()
 
@@ -198,9 +195,8 @@ describe('NewNorth', () => {
     })
 
     act(() => {
-      ReactDOM.render(
+      root.render(
         <NewNorth modal={false} toggle={() => false} />,
-        container,
       )
     })
     expect(container).toMatchSnapshot()
@@ -235,47 +231,82 @@ describe('NewNorth', () => {
     })
 
     act(() => {
-      ReactDOM.render(<EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />, container)
+      root.render(<EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />)
     })
-    Simulate.click(document.getElementById('dropdown-toggle'))
-    Simulate.click(document.getElementById('add-north'))
-    Simulate.click(document.getElementById('icon-connector'))
+    act(() => {
+      Simulate.click(document.getElementById('dropdown-toggle'))
+    })
+    act(() => {
+      Simulate.click(document.getElementById('add-north'))
+    })
+    act(() => {
+      Simulate.click(document.getElementById('icon-connector'))
+    })
     expect(document.getElementById('icon-connector').classList).toContain('connector-focus')
-    Simulate.change(document.getElementById('name'), { target: { value: 'new_application' } })
-    Simulate.click(document.getElementById('confirm'))
+    act(() => {
+      Simulate.change(document.getElementById('name'), { target: { value: 'new_application' } })
+    })
+    act(() => {
+      Simulate.click(document.getElementById('confirm'))
+    })
     expect(container).toMatchSnapshot()
   })
 
   test('check add connector without protocol & without new name', () => {
     act(() => {
-      ReactDOM.render(<EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />, container)
+      root.render(<EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />)
     })
-    Simulate.click(document.getElementById('dropdown-toggle'))
-    Simulate.click(document.getElementById('add-north'))
-    Simulate.click(document.getElementById('confirm'))
+    act(() => {
+      Simulate.click(document.getElementById('dropdown-toggle'))
+    })
+    act(() => {
+      Simulate.click(document.getElementById('add-north'))
+    })
+    act(() => {
+      Simulate.click(document.getElementById('confirm'))
+    })
     expect(container).toMatchSnapshot()
   })
 
   test('check add connector without protocol & with new name', () => {
     act(() => {
-      ReactDOM.render(<EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />, container)
+      root.render(<EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />)
     })
-    Simulate.click(document.getElementById('dropdown-toggle'))
-    Simulate.click(document.getElementById('add-north'))
-    Simulate.change(document.getElementById('name'), { target: { value: 'new_application' } })
+
+    act(() => {
+      Simulate.click(document.getElementById('dropdown-toggle'))
+    })
+    act(() => {
+      Simulate.click(document.getElementById('add-north'))
+    })
+    act(() => {
+      Simulate.change(document.getElementById('name'), { target: { value: 'new_application' } })
+    })
+    act(() => {
+      Simulate.click(document.getElementById('confirm'))
+    })
     expect(container).toMatchSnapshot()
-    Simulate.click(document.getElementById('confirm'))
   })
 
   test('check add connector without name & with protocol', () => {
     act(() => {
-      ReactDOM.render(<EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />, container)
+      root.render(<EngineMenu onRestart={restartFunction} onShutdown={shutdownFunction} />)
     })
-    Simulate.click(document.getElementById('dropdown-toggle'))
-    Simulate.click(document.getElementById('add-north'))
-    Simulate.click(document.getElementById('icon-connector'))
-    Simulate.change(document.getElementById('name'), { target: { value: '' } })
-    Simulate.click(document.getElementById('confirm'))
+    act(() => {
+      Simulate.click(document.getElementById('dropdown-toggle'))
+    })
+    act(() => {
+      Simulate.click(document.getElementById('add-north'))
+    })
+    act(() => {
+      Simulate.click(document.getElementById('icon-connector'))
+    })
+    act(() => {
+      Simulate.change(document.getElementById('name'), { target: { value: '' } })
+    })
+    act(() => {
+      Simulate.click(document.getElementById('confirm'))
+    })
     expect(container).toMatchSnapshot()
   })
 })
