@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import newConfig from '../../../tests/testConfig'
 import ConfigureApi from './ConfigureApi.jsx'
 import utils from '../helpers/utils'
@@ -23,28 +23,36 @@ mockMath.random = () => 1
 global.Math = mockMath
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('ConfigureApi', () => {
   test('check ConfigureApi', () => {
     act(() => {
-      ReactDOM.render(<ConfigureApi />, container)
+      root.render(<ConfigureApi />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check update', () => {
     act(() => {
-      ReactDOM.render(<ConfigureApi />, container)
+      root.render(<ConfigureApi />)
     })
-    Simulate.change(document.getElementById('north.applications.1.OIConnect.host'), { target: { value: 'http://new_host' } })
+    act(() => {
+      Simulate.change(document.getElementById('north.applications.1.OIConnect.host'), { target: { value: 'http://new_host' } })
+    })
+
     expect(dispatchNewConfig).toBeCalledWith({
       type: 'update',
       name: 'north.applications.1.OIConnect.host',
@@ -55,12 +63,12 @@ describe('ConfigureApi', () => {
   })
   test('check application not found', () => {
     const reactUseContextMock = React.useContext
-    // temporary empty applicartions array
+    // temporary empty applications array
     const config = utils.jsonCopy(newConfig)
     config.north.applications = []
     React.useContext = jest.fn().mockReturnValue({ newConfig: config, dispatchNewConfig })
     act(() => {
-      ReactDOM.render(<ConfigureApi />, container)
+      root.render(<ConfigureApi />)
     })
     expect(container).toMatchSnapshot()
     React.useContext = reactUseContextMock
