@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import newConfig from '../../../../tests/testConfig'
 import SubscribedTo from './SubscribedTo.jsx'
 
@@ -16,24 +16,29 @@ mockMath.random = () => 1
 global.Math = mockMath
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 newConfig.north.applications.forEach((application) => {
   describe('SubscribedTo', () => {
     test(`check SubscribedTo with application: ${application.name}`, () => {
       act(() => {
-        ReactDOM.render(<SubscribedTo
+        root.render(<SubscribedTo
           subscribedTo={application.subscribedTo}
           applicationIndex={0}
-        />, container)
+        />)
       })
       expect(container).toMatchSnapshot()
     })
@@ -44,12 +49,15 @@ const application = newConfig.north.applications[0]
 describe('SubscribedTo click changes', () => {
   test('check SubscribedTo add row', () => {
     act(() => {
-      ReactDOM.render(<SubscribedTo
+      root.render(<SubscribedTo
         subscribedTo={application.subscribedTo}
         applicationIndex={0}
-      />, container)
+      />)
     })
-    Simulate.click(document.querySelector('th path'))
+    act(() => {
+      Simulate.click(document.querySelector('th path'))
+    })
+
     expect(dispatchNewConfig).toBeCalledWith({
       type: 'addRow',
       name: 'north.applications.0.subscribedTo',
@@ -59,12 +67,15 @@ describe('SubscribedTo click changes', () => {
   })
   test('check SubscribedTo update row', () => {
     act(() => {
-      ReactDOM.render(<SubscribedTo
+      root.render(<SubscribedTo
         subscribedTo={application.subscribedTo}
         applicationIndex={0}
-      />, container)
+      />)
     })
-    Simulate.change(document.getElementById('subscribedTo.0'), { target: { value: 'MQTTServer', selectedIndex: 0 } })
+    act(() => (
+      Simulate.change(document.getElementById('subscribedTo.0'), { target: { value: 'MQTTServer', selectedIndex: 0 } })
+    ))
+
     expect(dispatchNewConfig).toBeCalledWith({
       type: 'update',
       name: 'north.applications.0.subscribedTo.0',
@@ -74,13 +85,18 @@ describe('SubscribedTo click changes', () => {
   })
   test('check SubscribedTo delete first row', () => {
     act(() => {
-      ReactDOM.render(<SubscribedTo
+      root.render(<SubscribedTo
         subscribedTo={application.subscribedTo}
         applicationIndex={0}
-      />, container)
+      />)
     })
-    Simulate.click(document.querySelector('td path'))
-    Simulate.click(document.getElementsByClassName('btn btn-primary')[0])
+    act(() => {
+      Simulate.click(document.querySelector('td path'))
+    })
+    act(() => {
+      Simulate.click(document.getElementsByClassName('btn btn-primary')[0])
+    })
+
     expect(dispatchNewConfig).toBeCalledWith({
       type: 'deleteRow',
       name: 'north.applications.0.subscribedTo.0',
@@ -93,10 +109,10 @@ describe('SubscribedTo no dataSources', () => {
   test('check SubscribedTo with no dataSources', () => {
     React.useContext = jest.fn().mockReturnValue({ dispatchNewConfig })
     act(() => {
-      ReactDOM.render(<SubscribedTo
+      root.render(<SubscribedTo
         subscribedTo={application.subscribedTo}
         applicationIndex={0}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
