@@ -2,8 +2,8 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
+import * as ReactDOMClient from 'react-dom/client'
 import newConfig from '../../../tests/testConfig'
 import NorthStatus from './NorthStatus.jsx'
 
@@ -28,28 +28,33 @@ mockMath.random = () => 1
 global.Math = mockMath
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('North Status', () => {
   React.useContext = jest.fn().mockReturnValue({ newConfig, dispatchNewConfig, setAlert })
   test('display NorthStatus page based on config', () => {
     act(() => {
-      ReactDOM.render(<NorthStatus />, container)
+      root.render(<NorthStatus />)
     })
     expect(container).toMatchSnapshot()
   })
 
   test('go to previous page', () => {
     act(() => {
-      ReactDOM.render(<NorthStatus />, container)
+      root.render(<NorthStatus />)
     })
     Simulate.click(document.getElementById('oi-sub-nav'))
     Simulate.click(document.getElementById('oi-navigate'))
