@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import testConfig from '../../../tests/testConfig'
 import PointsButton from './PointsButton.jsx'
 
@@ -14,14 +14,19 @@ jest.mock('react-router-dom', () => (
 ))
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 const dataSource = testConfig.south.dataSources[0]
@@ -34,26 +39,26 @@ describe('PointsButton', () => {
   test('check PointsButton disabled', () => {
     dataSource.enabled = false
     act(() => {
-      ReactDOM.render(<PointsButton
+      root.render(<PointsButton
         dataSource={dataSource}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check PointsButton enabled', () => {
     dataSource.enabled = true
     act(() => {
-      ReactDOM.render(<PointsButton
+      root.render(<PointsButton
         dataSource={dataSource}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check points button click', () => {
     act(() => {
-      ReactDOM.render(<PointsButton
+      root.render(<PointsButton
         dataSource={dataSource}
-      />, container)
+      />)
     })
     Simulate.click(document.querySelector('button.oi-points-button'))
     expect(mockNavigate).toBeCalledWith(`/south/${dataSource.id}/points`)
@@ -61,26 +66,26 @@ describe('PointsButton', () => {
   })
   test('check if points array is empty', () => {
     act(() => {
-      ReactDOM.render(<PointsButton
+      root.render(<PointsButton
         dataSource={emptyPointsDataSource}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check no points', () => {
     act(() => {
-      ReactDOM.render(<PointsButton
+      root.render(<PointsButton
         dataSource={nullPointsDataSource}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check if enabled with no points', () => {
     nullPointsDataSource.enabled = true
     act(() => {
-      ReactDOM.render(<PointsButton
+      root.render(<PointsButton
         dataSource={nullPointsDataSource}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
