@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import ConfirmationModal from './ConfirmationModal.jsx'
 
 import activeConfig from '../../../tests/testConfig'
@@ -16,25 +16,27 @@ jest.mock('react-router-dom', () => (
 ))
 
 let container
-
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
-  jest.clearAllMocks()
+  root = null
 })
 
 React.useContext = jest.fn().mockReturnValue({ activeConfig })
 describe('ConfirmationModal', () => {
   test('display ConfirmationModal page based on config', async () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <ConfirmationModal title="My Title" body="My Body" onConfirm={() => true} isOpen={false} toggle={() => true} />,
-        container,
       )
     })
     expect(container).toMatchSnapshot()
