@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import Modal from './Modal.jsx'
 
 const setOpen = jest.fn()
@@ -24,20 +24,25 @@ React.useState = jest.fn()
   })
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('Modal', () => {
   test('check Modal', () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <Modal
           show
           title="title"
@@ -47,7 +52,6 @@ describe('Modal', () => {
         >
           {() => (1)}
         </Modal>,
-        container,
       )
     })
     expect(container)
@@ -55,7 +59,7 @@ describe('Modal', () => {
   })
   test('check confirm press', () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <Modal
           show
           title="title"
@@ -65,7 +69,6 @@ describe('Modal', () => {
         >
           {() => (1)}
         </Modal>,
-        container,
       )
     })
     Simulate.click(document.getElementsByClassName('btn btn-primary')[0])
@@ -80,7 +83,7 @@ describe('Modal', () => {
   })
   test('check cancel press', () => {
     act(() => {
-      ReactDOM.render(
+      root.render(
         <Modal
           show
           title="title"
@@ -90,7 +93,6 @@ describe('Modal', () => {
         >
           {() => (1)}
         </Modal>,
-        container,
       )
     })
     Simulate.click(document.getElementsByClassName('btn btn-secondary')[0])
