@@ -2,74 +2,79 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import OIbAuthentication from './OIbAuthentication.jsx'
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('OIbAuthentication', () => {
   test('check Authentication with mode null', () => {
     act(() => {
-      ReactDOM.render(<OIbAuthentication
+      root.render(<OIbAuthentication
         value={{ type: 'Basic', username: 'user', password: 'pass' }}
         name="name"
         onChange={() => (1)}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check Authentication with mode API Key', () => {
     act(() => {
-      ReactDOM.render(<OIbAuthentication
+      root.render(<OIbAuthentication
         value={{ type: 'API Key', key: 'aaaaaa', secretKey: 'sssss' }}
         name="name"
         onChange={() => (1)}
         mode="API Key"
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check Authentication with mode Bearer', () => {
     act(() => {
-      ReactDOM.render(<OIbAuthentication
+      root.render(<OIbAuthentication
         value={{ type: 'Bearer', token: 'sssss' }}
         name="name"
         onChange={() => (1)}
         mode="Bearer"
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check Authentication with mode Basic', () => {
     act(() => {
-      ReactDOM.render(<OIbAuthentication
+      root.render(<OIbAuthentication
         value={{ type: 'Basic', username: 'user', password: 'pass' }}
         name="name"
         onChange={() => (1)}
         mode="Basic"
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check the returned value', () => {
     const onChangeMock = jest.fn()
     act(() => {
-      ReactDOM.render(<OIbAuthentication
+      root.render(<OIbAuthentication
         value={{ type: 'Basic', username: 'anyUser', password: 'anyPass' }}
         name="anyname"
         onChange={onChangeMock}
-      />, container)
+      />)
     })
 
     const userField = container.querySelector('#username')
@@ -83,11 +88,11 @@ describe('OIbAuthentication', () => {
   test('check the returned value (custom authentication)', () => {
     const onChangeMock = jest.fn()
     act(() => {
-      ReactDOM.render(<OIbAuthentication
+      root.render(<OIbAuthentication
         value={{ type: 'API Key', key: 'anyHeader', secretKey: 'anySecretKey' }}
         name="anyname"
         onChange={onChangeMock}
-      />, container)
+      />)
     })
 
     const keyField = container.querySelector('#key')
