@@ -2,60 +2,65 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import OIbInteger from './OIbInteger.jsx'
 import { minValue } from '../../../services/validation.service'
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('OIbInteger', () => {
   test('check OIbInteger with value="a text"', () => {
     act(() => {
-      ReactDOM.render(<OIbInteger
+      root.render(<OIbInteger
         label="label"
         value="12345678"
         name="name"
         onChange={() => (1)}
         defaultValue="0"
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check OIbInteger with too small value', () => {
     act(() => {
-      ReactDOM.render(<OIbInteger
+      root.render(<OIbInteger
         label="label"
         name="name"
         onChange={() => (1)}
         value="12"
         valid={minValue(100)}
         defaultValue="0"
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check change', () => {
     const onChange = jest.fn()
     act(() => {
-      ReactDOM.render(<OIbInteger
+      root.render(<OIbInteger
         label="label"
         name="name"
         onChange={onChange}
         value="12"
         valid={minValue(100)}
         defaultValue="0"
-      />, container)
+      />)
     })
     Simulate.change(document.querySelector('input'), { target: { value: 'new_value' } })
     expect(onChange).toBeCalled()
