@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import OIbScanMode from './OIbScanMode.jsx'
 
 jest.mock('react-router-dom', () => (
@@ -12,47 +12,52 @@ jest.mock('react-router-dom', () => (
 ))
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('OIbScanMode', () => {
   test('check ScanMode with value="value"', () => {
     act(() => {
-      ReactDOM.render(<OIbScanMode
+      root.render(<OIbScanMode
         name="name"
         label="label"
         value="value"
         onChange={() => (1)}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check ScanMode with no label', () => {
     act(() => {
-      ReactDOM.render(<OIbScanMode
+      root.render(<OIbScanMode
         name="name"
         value="value"
         onChange={() => (1)}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check ScanMode with value="value" and help', () => {
     act(() => {
-      ReactDOM.render(<OIbScanMode
+      root.render(<OIbScanMode
         name="name"
         help={<div>help text</div>}
         label="label"
         value="value"
         onChange={() => (1)}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
