@@ -2,9 +2,9 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act, Simulate } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import OIbTable from './OIbTable.jsx'
 
 const dispatchNewConfig = jest.fn()
@@ -15,46 +15,51 @@ jest.mock('react-router-dom', () => (
 ))
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('OIbTable', () => {
   test('check Table with name, label and rows', () => {
     act(() => {
-      ReactDOM.render(<OIbTable
+      root.render(<OIbTable
         name="a name"
         label="a title"
         rows={{ value: 'value' }}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check Table with help', () => {
     act(() => {
-      ReactDOM.render(<OIbTable
+      root.render(<OIbTable
         name="a name"
         label="a title"
         help={<div>help</div>}
         rows={{ value: 'value' }}
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
   test('check add to table', () => {
     act(() => {
-      ReactDOM.render(<OIbTable
+      root.render(<OIbTable
         name="name"
         label="a title"
         help={<div>help</div>}
         rows={{ value: 'value' }}
-      />, container)
+      />)
     })
     Simulate.click(document.querySelector('th path'))
     expect(dispatchNewConfig).toBeCalledWith({ type: 'addRow', name: 'name', value: {} })
@@ -62,16 +67,21 @@ describe('OIbTable', () => {
   })
   test('check delete from table', () => {
     act(() => {
-      ReactDOM.render(<OIbTable
+      root.render(<OIbTable
         name="name"
         label="a title"
         help={<div>help</div>}
         rows={{ scanMode: { type: 'OIbScanMode', label: 'Scan Mode' } }}
         value={[{ type: 'test', label: 'test' }]}
-      />, container)
+      />)
     })
-    Simulate.click(document.querySelector('td path'))
-    Simulate.click(document.getElementsByClassName('btn btn-primary')[0])
+    act(() => {
+      Simulate.click(document.querySelector('td path'))
+    })
+    act(() => {
+      Simulate.click(document.getElementsByClassName('btn btn-primary')[0])
+    })
+
     expect(dispatchNewConfig).toBeCalledWith({ type: 'deleteRow', name: 'name.0' })
     expect(container).toMatchSnapshot()
   })
