@@ -2,28 +2,33 @@
  * @jest-environment jsdom
  */
 import React from 'react'
-import ReactDOM from 'react-dom'
 import { act } from 'react-dom/test-utils'
 
+import * as ReactDOMClient from 'react-dom/client'
 import OIbTitle from './OIbTitle.jsx'
 
 let container
+let root
+// eslint-disable-next-line no-undef
+globalThis.IS_REACT_ACT_ENVIRONMENT = true
 beforeEach(() => {
   container = document.createElement('div')
+  root = ReactDOMClient.createRoot(container)
   document.body.appendChild(container)
 })
 
 afterEach(() => {
   document.body.removeChild(container)
   container = null
+  root = null
 })
 
 describe('OIbTitle', () => {
   test('check Title with no children', () => {
     act(() => {
-      ReactDOM.render(<OIbTitle
+      root.render(<OIbTitle
         label="a title"
-      />, container)
+      />)
     })
     expect(container).toMatchSnapshot()
   })
